# Improve OSX threading performance